+++ NEWS.md
@@ -1,6 +1,9 @@
 # Development version
 
 * Added tests for most functionality.
+* Reimplemented `truncate_alignment()` using `narrow()` method for Stockholm
+  alignments from `inferrnal` 0.99.8.  It now handles interleaved Stockholm
+  files.
 
 # LSUx 0.99.6